Move copying extension data tests
